Moved out of the __init__.py the functions: check_id_ok() and generate_random_pokemon()
